GUI_chatbot-trainer_TKinter/requirements.txt
# SCRIPT DE EJECUCION SERVER
# python src/scripts/dev_runner.py dev --port 3001

Pillow==9.2.0